+++ packages/import/client/Router.js
@@ -20,7 +20,7 @@ Router.map(function() {
 			return {
 				title: i18n('label_imported'),
 				className: 'imports',
-				headerTemplate: 'formImport',
+				headerTemplate: 'panelImport',
 				itemsTemplate: 'itemImport',
 				sortBy: 'name',
 				items: _.map(K.Profile.data.imports, function(i) {
@@ -46,7 +46,6 @@ Router.map(function() {
 			return {
 				title: i18n('label_imported')+': '+this.params.name,
 				className: 'imports-name',
-				//headerTemplate: 'formImportList',
 				itemsTemplate: 'itemPlaceImport',
 				items: _.map(K.findPlacesByImportName(K.Profile.id, this.params.name).fetch(), function(p) {
 					return K.placeById(p._id);